--- src/zkaf/apps/build.rs
@@ -1,2 +1,0 @@
-fn main() {
-}
